+++ test/fixtures/typescript/readonly.spec.ts
@@ -20,18 +20,16 @@ describe('typescript readonly tests', () => {
         },
         elements: [
           {
-            type: 'JsdocTypeKeyValue',
+            type: 'JsdocTypeObjectField',
             key: 'x',
+            readonly: true,
+            optional: false,
             right: {
               type: 'JsdocTypeName',
               value: 'number'
             },
-            optional: false,
-            readonly: true,
-            variadic: false,
             meta: {
-              quote: undefined,
-              hasLeftSideExpression: false
+              quote: undefined
             }
           }
         ]